Renamed new_syntax_test.py to fixture.py
